+++ collections/max_with.ts
@@ -21,7 +21,7 @@
  * ```
  */
 export function maxWith<T>(
-  array: readonly T[],
+  array: Iterable<T>,
   comparator: (a: T, b: T) => number,
 ): T | undefined {
   let max: T | undefined = undefined;