+++ collections/min_with.ts
@@ -17,7 +17,7 @@
  * ```
  */
 export function minWith<T>(
-  array: readonly T[],
+  array: Iterable<T>,
   comparator: (a: T, b: T) => number,
 ): T | undefined {
   let min: T | undefined = undefined;