+++ collections/partition.ts
@@ -19,15 +19,15 @@
  * ```
  */
 export function partition<T, U extends T>(
-  array: readonly T[],
+  array: Iterable<T>,
   predicate: (el: T) => el is U,
 ): [U[], Exclude<T, U>[]];
 export function partition<T>(
-  array: readonly T[],
+  array: Iterable<T>,
   predicate: (el: T) => boolean,
 ): [T[], T[]];
 export function partition(
-  array: readonly unknown[],
+  array: Iterable<unknown>,
   predicate: (el: unknown) => boolean,
 ): [unknown[], unknown[]] {
   const matches: Array<unknown> = [];