+++ http/file_server_test.ts
@@ -10,7 +10,14 @@ import { writeAll } from "../streams/write_all.ts";
 import { TextLineStream } from "../streams/text_line_stream.ts";
 import { serveDir, serveFile } from "./file_server.ts";
 import { calculate } from "./etag.ts";
-import { dirname, fromFileUrl, join, resolve, toFileUrl } from "../path/mod.ts";
+import {
+  basename,
+  dirname,
+  fromFileUrl,
+  join,
+  resolve,
+  toFileUrl,
+} from "../path/mod.ts";
 import { isWindows } from "../_util/os.ts";
 import { VERSION } from "../version.ts";
 import { retry } from "../async/retry.ts";
@@ -123,7 +130,7 @@ async function fetchExactPath(
   const encoder = new TextEncoder();
   const decoder = new TextDecoder();
   const request = encoder.encode("GET " + path + " HTTP/1.1\r\n\r\n");
-  let conn: void | Deno.Conn;
+  let conn: undefined | Deno.Conn;
   try {
     conn = await Deno.connect(
       { hostname: hostname, port: port, transport: "tcp" },
@@ -517,32 +524,45 @@ async function startTlsFileServer({
   const res = await reader.read();
   assert(!res.done && res.value.includes("Listening"));
   reader.releaseLock();
-}
 
-Deno.test("serveDirIndex TLS", async function () {
-  await startTlsFileServer();
-  try {
-    // Valid request after invalid
+  // Wait for fileServer to be ready.
+  await retry(async () => {
     const conn = await Deno.connectTls({
       hostname: "localhost",
-      port: 4577,
+      port: +port,
       certFile: join(testdataDir, "tls/RootCA.pem"),
     });
-
-    await writeAll(
-      conn,
-      new TextEncoder().encode("GET / HTTP/1.0\r\n\r\n"),
-    );
-    const res = new Uint8Array(128 * 1024);
-    const nread = await conn.read(res);
-    assert(nread !== null);
     conn.close();
-    const page = new TextDecoder().decode(res.subarray(0, nread));
-    assert(page.includes("<title>Deno File Server</title>"));
-  } finally {
-    await killFileServer();
-  }
-});
+  });
+}
+
+Deno.test(
+  "serveDirIndex TLS",
+  async function () {
+    await startTlsFileServer();
+    try {
+      // Valid request after invalid
+      const conn = await Deno.connectTls({
+        hostname: "localhost",
+        port: 4577,
+        certFile: join(testdataDir, "tls/RootCA.pem"),
+      });
+
+      await writeAll(
+        conn,
+        new TextEncoder().encode("GET / HTTP/1.0\r\n\r\n"),
+      );
+      const res = new Uint8Array(128 * 1024);
+      const nread = await conn.read(res);
+      assert(nread !== null);
+      conn.close();
+      const page = new TextDecoder().decode(res.subarray(0, nread));
+      assert(page.includes("<title>Deno File Server</title>"));
+    } finally {
+      await killFileServer();
+    }
+  },
+);
 
 Deno.test("partial TLS arguments fail", async function () {
   const fileServer = new Deno.Command(Deno.execPath(), {
@@ -1530,6 +1550,79 @@ Deno.test(
       }
     } finally {
       await killFileServer();
+    }
+  },
+);
+
+Deno.test("file_server should resolve `path` correctly on Windows", {
+  ignore: Deno.build.os !== "windows",
+}, async () => {
+  const fileServer = new Deno.Command(Deno.execPath(), {
+    args: [
+      "run",
+      "--no-check",
+      "--quiet",
+      "--allow-read",
+      "--allow-net",
+      "file_server.ts",
+      "c:/",
+      "--host",
+      "localhost",
+      "--port",
+      `4507`,
+    ],
+    cwd: moduleDir,
+    stdout: "null",
+    stderr: "null",
+  });
+  child = fileServer.spawn();
+  try {
+    const resp = await fetch("http://localhost:4507/");
+    assertEquals(resp.status, 200);
+    await resp.text(); // Consuming the body so that the test doesn't leak resources
+  } finally {
+    await killFileServer();
+  }
+});
+
+Deno.test(
+  "file_server should resolve empty subdir correctly without asking for current directory read permission on Windows",
+  {
+    ignore: Deno.build.os !== "windows",
+  },
+  async () => {
+    const tempDir = Deno.makeTempDirSync({ dir: `${moduleDir}/testdata` });
+    const fileServer = new Deno.Command(Deno.execPath(), {
+      // specifying a path for `--allow-read` this is essential for this test
+      // otherwise it won't trigger the edge case
+      args: [
+        "run",
+        "--no-check",
+        "--no-prompt",
+        "--quiet",
+        `--allow-read=${moduleDir}/testdata`,
+        "--allow-net",
+        "file_server.ts",
+        moduleDir,
+        "--host",
+        "localhost",
+        "--port",
+        "4507",
+      ],
+      cwd: moduleDir,
+      stdout: "null",
+      stderr: "null",
+    });
+    child = fileServer.spawn();
+    try {
+      const resp = await fetch(
+        `http://localhost:4507/testdata/${basename(tempDir)}`,
+      );
+      assertEquals(resp.status, 200);
+      await resp.text(); // Consuming the body so that the test doesn't leak resources
+    } finally {
+      await killFileServer();
+      Deno.removeSync(tempDir);
     }
   },
 );