+++ http/util.ts
@@ -4,23 +4,6 @@
 import { Status, STATUS_TEXT } from "./http_status.ts";
 import { deepMerge } from "../collections/deep_merge.ts";
 
-/** Returns true if the etags match. Weak etag comparisons are handled.
- *
- * @deprecated (will be removed after 0.187.0) use `etag/ifMatch` and `etag/ifNoneMatch` instead.
- */
-export function compareEtag(a: string, b: string): boolean {
-  if (a === b) {
-    return true;
-  }
-  if (a.startsWith("W/") && !b.startsWith("W/")) {
-    return a.slice(2) === b;
-  }
-  if (!a.startsWith("W/") && b.startsWith("W/")) {
-    return a === b.slice(2);
-  }
-  return false;
-}
-
 /**
  * Internal utility for returning a standardized response, automatically defining the body, status code and status text, according to the response type.
  */