+++ toml/test.ts
@@ -398,7 +398,7 @@ Deno.test({
       foo: { bar: "deno" },
       this: { is: { nested: "denonono" } },
       "https://deno.land/std": {
-        $: "doller",
+        $: "dollar",
       },
       "##": {
         deno: {
@@ -491,7 +491,7 @@ bar = "deno"
 nested = "denonono"
 
 ["https://deno.land/std"]
-"$" = "doller"
+"$" = "dollar"
 
 ["##".deno."https://deno.land"]
 proto = "https"
@@ -708,6 +708,52 @@ aaa   = 1
 aaaa  = 1
 aaaaa = 1
 `;
+    assertEquals(actual, expected);
+  },
+});
+
+Deno.test({
+  name: "[TOML] stringify empty key",
+  fn() {
+    const src = {
+      "": "a",
+      "b": { "": "c" },
+    };
+    const actual = stringify(src);
+    const expected = `"" = "a"
+
+[b]
+"" = "c"
+`;
+    assertEquals(actual, expected);
+  },
+});
+
+Deno.test({
+  name: "[TOML] stringify empty object",
+  fn() {
+    const src = {
+      "a": {},
+      "b": { "c": {} },
+    };
+    const actual = stringify(src);
+    const expected = `
+[a]
+
+[b.c]
+`;
+    assertEquals(actual, expected);
+  },
+});
+
+Deno.test({
+  name: "[TOML] stringify special keys in inline object",
+  fn() {
+    const src = {
+      "a": [{ "/": "b" }, "c"],
+    };
+    const actual = stringify(src);
+    const expected = 'a = [{"/" = "b"},"c"]\n';
     assertEquals(actual, expected);
   },
 });