+++ Releases.md
@@ -1,3 +1,45 @@
+### 0.193.0 / 2023.07.04
+
+- feat: add http/user_agent (#3387)
+- fix(json): allow primitives at top level without separator (#3466)
+- fix(testing/asserts): handle primitive/null values better in assertObjectMatch
+  (#3468)
+
+### 0.192.0 / 2023.06.15
+
+- BREAKING(semver): rewrite semver (#3385)
+- feat(testing): report the number of removed snapshots (#3435)
+- fix(datetime/day_of_year): respect time zone of date (#3443)
+- fix(http/file_server): resolve empty subdir correctly on Windows (#3439)
+- fix(testing/time): use real Date in FakeTime (#3414)
+- fix(yaml): parse always return null when file is empty, whitespace or only
+  comments (#3442)
+
+### 0.191.0 / 2023.06.08
+
+- BREAKING(csv,http,path): remove deprecated APIs (#3418)
+- feat(async/retry): introduce jitter option and fix retry bugs (#3427)
+- feat(collections): switch functions to take iterables when possible (#3401)
+- feat(collections/sort_by): descending order can be specified in options
+  (#3419)
+- feat(crypto): add BLAKE2B-224 hashing algorithm in crypto (#3392)
+- feat(html): add escape and unescape functions for HTML entities (#3335)
+- fix(http/file_server): use platform specific `resolve` (#3424)
+- fix(streams/mergeReadableStreams): better error handling (#3395)
+- fix(toml): various edge case fixes for `toml.stringify` (#3403)
+
+### 0.190.0 / 2023.05.29
+
+### 0.189.0 / 2023.05.24
+
+- feat(async): add jitter to retry exponential backoff (#3379)
+- feat(collections/group_by): accept iterable input, add index param to callback
+  (#3390)
+
+### 0.188.0 / 2023.05.18
+
+- fix(flags): correctly collect default value (#3380)
+
 ### 0.187.0 / 2023.05.12
 
 - feat(collections): add `partitionEntries` (#3365)
@@ -121,7 +163,7 @@
 ### 0.176.0 / 2023.02.02
 
 - fix(node): disable worker_threads (#3151)
-- fix(node): throw permisison error instead of unknown error (#3133)
+- fix(node): throw permission error instead of unknown error (#3133)
 - fix(node/util): stricter runtime type checking (#3122)
 - fix: make encoding/front_matter work in a browser (#3154)
 
@@ -868,7 +910,7 @@ new feature added setNoDelay.
 - feat(node): allow require with 'node:' prefix (#1438)
 - feat(node/url): add `url.urlToHttpOptions(url)` (#1426)
 - feat(testing): add assertIsError (#1376)
-- fix(async): fix async/tee concurent .next calls error (#1425)
+- fix(async): fix async/tee concurrent .next calls error (#1425)
 - fix(crypto): support length option in crypto.subtle.digest (#1386)
 - fix(http/file_server): fix encoded url in dir html (#1442)
 - fix(http/file_server): fix leak file resource (#1443)
@@ -1148,7 +1190,7 @@ new feature added setNoDelay.
 ### 0.92.0 / 2021.04.02
 
 - feat: make bufio compatible to Deno Deploy (#831)
-- feat: add symlink adn symlinkSync to node/fs (#825)
+- feat: add symlink and symlinkSync to node/fs (#825)
 - feat: add format and improve deprecate in node/util (#693)
 - feat: add io/buffer and io/util module (#808) …
 - fix: handle upstream type changes (#834)